requirements.txt
pillow>=8.0.1
sv-ttk